refactor: extract metadata generation tab logic into separate ViewModel from MainWindowViewModel
